requirements.txt
alembic==1.16.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==22.0.0
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2024.2
Werkzeug==3.0.4
WTForms==3.1.2
xlsxwriter
requests==2.32.3
pytest==8.3.5
pytest-flask==1.3.0
selenium==4.25.0
